Fixes hitboxes and adds save/load
I am sorry this has taken me so long. There were enough students having trouble, that I decided to offer a solution to everyone. You still have a money-pump problem (you could save which coins are missing, for example), but this is at least a beginning. I hope this helps.